+++ Venmo/src/test/java/com/braintreepayments/api/VenmoRequestUnitTest.java
@@ -68,6 +68,16 @@ public class VenmoRequestUnitTest {
     }
 
     @Test
+    public void getIsFinalAmountAsString_returnsStringEquivalent() {
+        VenmoRequest request = new VenmoRequest(VenmoPaymentMethodUsage.MULTI_USE);
+        request.setIsFinalAmount(true);
+        assertEquals("true", request.getIsFinalAmountAsString());
+
+        request.setIsFinalAmount(false);
+        assertEquals("false", request.getIsFinalAmountAsString());
+    }
+
+    @Test
     public void parcelsCorrectly() {
         VenmoRequest request = new VenmoRequest(VenmoPaymentMethodUsage.MULTI_USE);
         request.setDisplayName("venmo-user");
@@ -80,6 +90,7 @@ public class VenmoRequestUnitTest {
         request.setDiscountAmount("2.00");
         request.setShippingAmount("1.00");
         request.setTotalAmount("10.00");
+        request.setIsFinalAmount(true);
         request.setFallbackToWeb(true);
 
         ArrayList<VenmoLineItem> lineItems = new ArrayList<>();
@@ -104,6 +115,7 @@ public class VenmoRequestUnitTest {
         assertEquals("10.00", result.getTotalAmount());
         assertEquals(1, result.getLineItems().size());
         assertEquals("An Item", result.getLineItems().get(0).getName());
+        assertTrue(result.getIsFinalAmount());
         assertEquals(true, result.getFallbackToWeb());
     }
 }